Remove the use of "part" and "part of"
The use of parts hides the dependency structure between
files.  Moreover, it's not a pattern that's often used
by the Dart code team, which means its doesn't get tested
as much.

This change switches the image package away from the use
of parts in favor of the more modern patern of exporting
the public API from the main lib file.