Add physical pixel size handling for PNG.
This implements support for the 'pHYs' chunk in the standard PNG format
specification, controlling the intended physical pixel size. The
physical pixel size (often specified in DPI) influences the displayed
size of the image for viewers that implement it.

One usecase for this functionality is encoding HiDPI (e.g. 2x, 4x)
images for display on Apple Retina displays.